README.md
This is a repository that contains the TCP code for creating and handling multiple connections to a server. With the server open, any number of clients can connect to the server given the correct IP address and port number. This is written specifically in conjunction with the ULI Tech Demo. It is currently written to only send the last  line of the flightlog data. The server also takes the transferred data and writes it to a file on the server machine.